refactor(#1): naming
